fix Consignment

+++ src/main/java/com/example/koifarm/model/KoiRequest.java
@@ -28,7 +28,7 @@ public class KoiRequest {
 
     @NotNull(message = "Size is required")
     @DecimalMin(value = "0.0", inclusive = false, message = "Size must be greater than 0")
-    private BigDecimal size;
+    private float size;
 
     @DecimalMin(value = "0.01", inclusive = false, message = "Price must be greater than 0")
     @DecimalMax(value = "99999999.99", message = "Price must be less than or equal to 99999999.99")
